+++ tests/k6/tests/serviceowner/performance/purge-dialogs.js
@@ -15,6 +15,8 @@ import { getDefaultThresholds } from '../../performancetest_common/getDefaultThr
 import { describe } from '../../../common/describe.js';
 import { sentinelPerformanceValue as sentinelValue } from '../../../common/config.js';
 
+const traceCalls = (__ENV.traceCalls ?? 'false') === 'true';
+
 /**
  * Retrieves the dialog ids to purge.
  * 
@@ -23,18 +25,24 @@ import { sentinelPerformanceValue as sentinelValue } from '../../../common/confi
  */
 function getDialogs(serviceOwner) {
     var traceparent = uuidv4();
-    console.log("Searching for dialogs to purge, tracevalue: " + traceparent);
     var paramsWithToken = {
         headers: {
             Authorization: "Bearer " + serviceOwner.token,
             traceparent: traceparent
-        }
+        },
+        tags: { name: 'search dialogs' }
     }
+
     let hasNextPage = false;
     let continuationToken = "";
     let dialogIdsToPurge = [];
     do {
-        let r = getSO('dialogs/?Search=' + sentinelValue + continuationToken, paramsWithToken);
+        traceparent = uuidv4();
+        paramsWithToken.headers.traceparent = traceparent;
+        if (traceCalls) {
+            paramsWithToken.tags.traceparent = traceparent;
+        }
+        let r = getSO('dialogs/?Search=' + encodeURIComponent(sentinelValue) + continuationToken, paramsWithToken);
         expectStatusFor(r).to.equal(200);
         expect(r, 'response').to.have.validJsonBody();
         let response = r.json();
@@ -49,12 +57,7 @@ function getDialogs(serviceOwner) {
             }
         }
         console.log("Found " + dialogIdsToPurge.length + " unpurged dialogs");  
-        if (dialogIdsToPurge.length < 200) {
-            hasNextPage = response.hasNextPage;
-        }
-        else {
-            hasNextPage = false;
-        }
+        hasNextPage = response.hasNextPage;
     } while (hasNextPage);
     return dialogIdsToPurge;
 }
@@ -107,17 +110,25 @@ export default function(serviceOwners) {
  * @param {Object} serviceOwner - The service owner object.
  */
 export function purgeDialogs(serviceOwner) {
+    var traceparent = uuidv4();
     var paramsWithToken = {
         headers: {
-            Authorization: "Bearer " + serviceOwner.token
-        },
-        tags: { name: 'purge dialog' }
+            Authorization: "Bearer " + serviceOwner.token,
+            traceparent: traceparent
+            },
+        tags: { name: 'purge dialog'}
     }
+
     describe('Post run: checking for unpurged dialogs', () => {
         let dialogIdsToPurge = serviceOwner.dialogIdsToPurge;
         if (dialogIdsToPurge.length > 0) {
             console.error("Found " + dialogIdsToPurge.length + " unpurged dialogs, make sure that all tests clean up after themselves. Purging ...");
             for(var i = dialogIdsToPurge.length - 1; i>=0; i--) {
+                traceparent = uuidv4();
+                paramsWithToken.headers.traceparent = traceparent;
+                if (traceCalls) {
+                    paramsWithToken.tags.traceparent = traceparent;
+                }
                 let r = purgeSO('dialogs/' + dialogIdsToPurge[i], paramsWithToken);
                 if (r.status != 204) {
                     console.error("Failed to purge dialog with id: " + dialogIdsToPurge[i]);